JAMES-2589 Solve concurency problem in vacation mailet on top of RabbitMQ mail queue
Now sending a mail requires some cassandra joins.

A join was called from within the future chain every time a notification was sent, leading to
Cassandra driver error.

Making the VacationMailet code synchronous solved the issue.

(Another alternative would be to "futurify" the MailQueue API + the MailetContext API but for sure it is more expensive.