MAILBOX-371 Remove unnecessary test for lifecycle
This test is in SingleLifeCycleTest, There is another equivalent test
with multiple event bus. Remove this because it is impossible to passed
when RabbitMQEventBusTest also implements MultipleLifeCycleTest, then in
setup phrase, two eventbus are up, that means, stopping one eventbus
doesn't make queue registration for key deleted. (Because there is another
eventbus is consuming this queue!)